tools/testing/connection/python-postgres/fastapi_integration_poc/requirements.txt
fastapi
uvicorn
pydantic[email]
python-dotenv
# Database dependencies
sqlalchemy==2.0.41
sqlalchemy[asyncio]==2.0.41
alembic>=1.13.1
asyncpg>=0.30.0
psycopg2-binary==2.9.9
